+++ src/talismans/example.cpp
@@ -1,0 +1,18 @@
+#include "example.h"
+
+
+
+Example::Example()
+{
+    a = 0;
+}
+
+void Example::process()
+{
+    a += 1;
+}
+
+const int Example::getResult()
+{
+    return a;
+}