+++ src/talismans/example.h
@@ -1,0 +1,15 @@
+// example.h
+
+class Example
+{
+private:
+    int a;
+
+public:
+    Example();
+
+    void process();
+
+    const int getResult();
+
+};