+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added wrapping of a C++ class ([c27a19a](ssh://git@git:7999/~blaudiy/talismans/commit/c27a19aad097d8f4deb6005e4747537420f29df1) by Yann BLAUDIN DE THE).
 - Added tests ([bc9af56](ssh://git@git:7999/~blaudiy/talismans/commit/bc9af565640370e954ff3db4e6998c8e8e241920) by Yann BLAUDIN DE THE).
 - Added CHANGELOG ([1560b2c](ssh://git@git:7999/~blaudiy/talismans/commit/1560b2c051c6562b9059dbb0424d166641fbf9e8) by Yann BLAUDIN DE THE).